requirements.txt
amqp==5.3.1
asgiref==3.8.1
babel==2.17.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.19
django-cors-headers==4.7.0
djangorestframework==3.15.2
docxcompose==1.4.0
docxtpl==0.19.1
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.5
kombu==5.4.2
lxml==5.3.0
MarkupSafe==3.0.2
pillow==11.1.0
prompt_toolkit==3.0.50
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
